Fix partially written SVGs.
...and add to the docs.

My emacs24 cut the output after the last match of the
replace-string call in `osm-publish-map-for-html'.  I
could not reproduce the bug in `emacs -Q' though.
Fixed now anyway.

I'm still trying to avoid the image-mode on export, but
to no avail.
  `Type C-c C-c to view the image as text'
is really annoying (this happens in `emacs -Q', too).